#28 - 소셜 회원 탈퇴 관련 클래스 정리
- 모든 소셜 회원 탈퇴 관련 기능은 RevokeController, RevokeService에서 처리하도록 함
