+++ src/main/java/com/ncookie/imad/global/oauth2/controller/AppleController.java
@@ -3,22 +3,17 @@ package com.ncookie.imad.global.oauth2.controller;
 import com.ncookie.imad.global.dto.response.ApiResponse;
 import com.ncookie.imad.global.dto.response.ResponseCode;
 import com.ncookie.imad.global.oauth2.service.AppleService;
-import com.ncookie.imad.global.oauth2.service.RevokeService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.io.IOException;
 
 @RestController
 @RequiredArgsConstructor
 public class AppleController {
     private final AppleService appleService;
-    private final RevokeService revokeService;
 
     @PostMapping("/api/callback/apple")
     public ApiResponse<?> callback(HttpServletRequest request, HttpServletResponse response) {
@@ -28,11 +23,5 @@ public class AppleController {
         }
 
         return ApiResponse.createSuccessWithNoContent(ResponseCode.LOGIN_SUCCESS);
-    }
-
-    @DeleteMapping("/api/oauth2/revoke/apple")
-    public ApiResponse<?> revokeAppleAccount(@RequestHeader("Authorization") String accessToken) throws IOException {
-        revokeService.deleteAppleAccount(accessToken);
-        return ApiResponse.createSuccessWithNoContent(ResponseCode.USER_DELETE_SUCCESS);
     }
 }
